Seperate cusparse util functions, improve robustness
Transferred the cusparse matrix product util functions to src/seq_mv/csr_matrix_cuda_utils.{h,c}
Improved the robustness by dynamically adapting to the HYPRE_Int/HYPRE_Complex data type
