+++ src/seq_mv/Makefile
@@ -37,6 +37,7 @@ FILES =\
 
 CUFILES =\
  csr_matop_device.c\
+ csr_matrix_cuda_utils.c\
  csr_matvec_device.c\
  csr_matvec_oomp.c\
  csr_spadd_device.c\
